+++ modules/dice-calculator/module/dice-calculator.js
@@ -545,6 +545,7 @@ class TemplateDiceMap {
 			this._extraButtonsLogic(html);
 		}
 
+		if (html.is("aside")) return;
 		/** Clicking the Roll button clears and hides all orange number flags, and unmark the KH/KL keys */
 		html.find(".dice-tray__roll").on("click", (event) => {
 			event.preventDefault();
@@ -596,6 +597,7 @@ class TemplateDiceMap {
 			event.preventDefault();
 			let dataset = event.currentTarget.dataset;
 			let $chat = html.find("#chat-form textarea");
+			if (html.is("aside")) $chat = this.popout.chatElement.find("#chat-form textarea");
 			let chat_val = String($chat.val());
 			let match_string = /\d*d\d+[khl]*/;
 
@@ -654,7 +656,10 @@ class TemplateDiceMap {
 	 * @param {HTMLElement} html
 	 */
 	applyModifier(html) {
-		const $mod_input = html.find(".dice-tray__input");
+		let $mod_input = html.find(".dice-tray__input");
+		if (!$mod_input.length && this.popout?.rendered) {
+			$mod_input = $(this.popout.element).find(".dice-tray__input");
+		}
 		const mod_val = Number($mod_input.val());
 
 		if ($mod_input.length === 0 || isNaN(mod_val)) return;
@@ -796,19 +801,24 @@ class TemplateDiceMap {
 
 		// Add a flag indicator on the dice.
 		qty = Number(qty);
-		const $flag_button = html.find(`.dice-tray__flag--${dataset.formula}`);
+		let $flag_button = [html.find(`.dice-tray__flag--${dataset.formula}`)];
+		if (this.popout?.rendered) {
+			$flag_button.push($(this.popout.element).find(`.dice-tray__flag--${dataset.formula}`));
+		}
 		if (!qty) {
 			qty = direction === "add" ? 1 : 0;
 		}
 
-		if (qty > 0) {
-			$flag_button.text(qty);
-			$flag_button.removeClass("hide");
-		} else if (qty < 0) {
-			$flag_button.text(qty);
-		} else {
-			$flag_button.text("");
-			$flag_button.addClass("hide");
+		for (const button of $flag_button) {
+			if (qty > 0) {
+				button.text(qty);
+				button.removeClass("hide");
+			} else if (qty < 0) {
+				button.text(qty);
+			} else {
+				button.text("");
+				button.addClass("hide");
+			}
 		}
 
 		currFormula = $chat.val();
@@ -1436,6 +1446,171 @@ class DiceCalculatorDialog extends Dialog {
 	}
 }
 
+const { ApplicationV2, HandlebarsApplicationMixin } = foundry.applications.api;
+
+class DiceTrayPopOut extends HandlebarsApplicationMixin(ApplicationV2) {
+	static DEFAULT_OPTIONS = {
+		id: "dice-tray-popout",
+		tag: "aside",
+		position: {
+			width: ui?.sidebar?.options.width ?? 300
+		},
+		window: {
+			title: "DICE_TRAY.DiceTray",
+			icon: "fas fa-dice-d20",
+			minimizable: true
+		}
+	};
+
+	static PARTS = {
+		list: {
+			id: "list",
+			template: "modules/dice-calculator/templates/tray.html",
+		}
+	};
+
+	async _renderFrame(options) {
+		const frame = await super._renderFrame(options);
+		this.window.close.remove(); // Prevent closing
+		return frame;
+	}
+
+	async close(options={}) {
+		if ( !options.closeKey ) return super.close(options);
+		return this;
+	}
+
+	get chatElement() {
+		return ui.sidebar.popouts.chat?.element || ui.chat.element;
+	}
+
+	_onFirstRender(context, options) {
+		super._onFirstRender(context, options);
+		const position = game.settings.get("dice-calculator", "popoutPosition");
+		const left = position.left ?? ui.nav?.element[0].getBoundingClientRect().left;
+		const top = position.top ?? ui.controls?.element[0].getBoundingClientRect().top;
+		options.position = {...options.position, left, top};
+		canvas.scene.apps[this.id] = this;
+	}
+
+	_onRender(context, options) {
+		super._onRender(context, options);
+		for (const section of this.element.querySelectorAll(".dice-tray__button")) {
+			section.addEventListener("click", (event) => {
+				event.preventDefault();
+				let dataset = event.currentTarget.dataset;
+
+				CONFIG.DICETRAY.updateChatDice(dataset, "add", this.chatElement);
+			});
+			section.addEventListener("contextmenu", (event) => {
+				event.preventDefault();
+				let dataset = event.currentTarget.dataset;
+
+				CONFIG.DICETRAY.updateChatDice(dataset, "sub", this.chatElement);
+			});
+		}
+		for (const section of this.element.querySelectorAll(".dice-tray__button, .dice-tray__ad")) {
+			section.addEventListener("dragstart", (event) => {
+				const dataset = event.currentTarget.dataset;
+				const dragData = JSON.parse(JSON.stringify(dataset));
+				if (dragData?.formula) {
+					// Grab the modifier, if any.
+					const $parent = $(event.currentTarget).parents(".dice-tray");
+					const $modInput = $parent.find(".dice-tray__input");
+					const mod = $modInput.val();
+
+					// Handle advantage/disadvantage.
+					if (dragData.formula === "kh") {
+						dragData.formula = "2d20kh";
+					} else if (dragData.formula === "kl") {
+						dragData.formula = "2d20kl";
+					}
+
+					// Grab the count, if any.
+					const qty = $(event.currentTarget).find(".dice-tray__flag").text();
+					if (qty.length > 0 && !dragData.formula.includes("k")) {
+						dragData.formula = `${qty}${dataset.formula}`;
+					}
+
+					// Apply the modifier.
+					if (mod && mod !== "0") {
+						dragData.formula += ` + ${mod}`;
+					}
+					dragData.origin = "dice-calculator";
+					event.originalEvent.dataTransfer.setData("text/plain", JSON.stringify(dragData));
+				}
+			});
+		}
+		for (const section of this.element.querySelectorAll(".dice-tray__input")) {
+			section.addEventListener("input", (event) => {
+				// event.preventDefault();
+				let $self = $(event.currentTarget);
+				let mod_val = $self.val();
+
+				mod_val = Number(mod_val);
+				mod_val = Number.isNaN(mod_val) ? 0 : mod_val;
+
+				$self.val(mod_val);
+				CONFIG.DICETRAY.applyModifier(this.chatElement);
+			});
+			section.addEventListener("wheel", (event) => {
+				let $self = $(event.currentTarget);
+				let diff = event.originalEvent.deltaY < 0 ? 1 : -1;
+				let mod_val = $self.val();
+				mod_val = Number.isNaN(mod_val) ? 0 : Number(mod_val);
+				$self.val(mod_val + diff);
+				CONFIG.DICETRAY.applyModifier(this.chatElement);
+			});
+		}
+		for (const section of this.element.querySelectorAll("button.dice-tray__math")) {
+			section.addEventListener("click", (event) => {
+				event.preventDefault();
+				let dataset = event.currentTarget.dataset;
+				let mod_val = $('input[name="dice.tray.modifier"]').val();
+
+				mod_val = Number(mod_val);
+				mod_val = Number.isNaN(mod_val) ? 0 : mod_val;
+
+				switch (dataset.formula) {
+					case "+1":
+						mod_val = mod_val + 1;
+						break;
+					case "-1":
+						mod_val = mod_val - 1;
+						break;
+				}
+				$('input[name="dice.tray.modifier"]').val(mod_val);
+				CONFIG.DICETRAY.applyModifier(this.chatElement);
+			});
+		}
+		CONFIG.DICETRAY.applyLayout($(this.element));
+		this.element.querySelector(".dice-tray__roll").addEventListener("click", (event) => {
+			event.preventDefault();
+			let spoofed = CONFIG.DICETRAY.triggerRollClick();
+			this.chatElement.find("#chat-message").trigger(spoofed);
+			CONFIG.DICETRAY._resetTray($(this.element));
+		});
+		this.chatElement.find("#chat-message").keydown((e) => {
+			if (e.code === "Enter" || e.key === "Enter" || e.keycode === "13") {
+				CONFIG.DICETRAY._resetTray($(this.element));
+			}
+		});
+	}
+
+	async _prepareContext(_options) {
+		return {
+			dicerows: game.settings.get("dice-calculator", "diceRows"),
+		};
+	}
+
+	setPosition(position) {
+		const superPosition = super.setPosition(position);
+		const { left, top } = superPosition;
+		game.settings.set("dice-calculator", "popoutPosition", { left, top });
+		return superPosition;
+	}
+}
+
 /**
  * Javascript imports don't support dashes so the workaround
  * for systems with dashes in their names is to create this map.
@@ -1799,6 +1974,37 @@ function registerSettings() {
 		default: [],
 		type: Array,
 	});
+
+	game.settings.register("dice-calculator", "popout", {
+		name: "DICE_TRAY.SETTINGS.popout.name",
+		hint: "DICE_TRAY.SETTINGS.popout.hint",
+		scope: "client",
+		config: true,
+		default: "none",
+		choices: {
+			none: "",
+			tokens: game.i18n.localize("CONTROLS.GroupToken"),
+			all: game.i18n.localize("DICE_TRAY.SETTINGS.popout.options.all"),
+		},
+		type: String,
+		onChange: () => ui.controls.initialize()
+	});
+
+	game.settings.register("dice-calculator", "autoOpenPopout", {
+		name: "DICE_TRAY.SETTINGS.autoOpenPopout.name",
+		hint: "DICE_TRAY.SETTINGS.autoOpenPopout.hint",
+		scope: "client",
+		config: true,
+		default: false,
+		type: Boolean
+	});
+
+	game.settings.register("dice-calculator", "popoutPosition", {
+		scope: "client",
+		config: false,
+		default: {},
+		type: Object
+	});
 }
 
 // Initialize module
@@ -1826,6 +2032,22 @@ Hooks.once("i18nInit", () => {
 	}
 
 	registerSettings();
+	game.keybindings.register("dice-calculator", "popout", {
+		name: "DICE_TRAY.KEYBINGINDS.popout.name",
+		onDown: async () => {
+			await togglePopout();
+			if (game.settings.get("dice-calculator", "popout") === "none") return;
+			const tool = ui.controls.control.tools.find((t) => t.name === "dice-tray");
+			if (tool) {
+				tool.active = !tool.active;
+				ui.controls.render();
+			}
+		}
+	});
+});
+
+Hooks.once("ready", () => {
+	if (game.settings.get("dice-calculator", "autoOpenPopout")) togglePopout();
 });
 
 function getProviderString(regex) {
@@ -1836,6 +2058,12 @@ function getProviderString(regex) {
 		return id;
 	}
 	return "";
+}
+
+async function togglePopout() {
+	CONFIG.DICETRAY.popout ??= new DiceTrayPopOut();
+	if (CONFIG.DICETRAY.popout.rendered) await CONFIG.DICETRAY.popout.close({ animate: false });
+	else await CONFIG.DICETRAY.popout.render(true);
 }
 
 Hooks.on("renderSidebarTab", async (app, html, data) => {
@@ -1869,7 +2097,6 @@ Hooks.on("renderSidebarTab", async (app, html, data) => {
 				CONFIG.DICETRAY.updateChatDice(dataset, "sub", html);
 			});
 			// Handle drag events.
-			$content.find(".dice-tray__button").attr("draggable", true);
 			$content.on("dragstart", ".dice-tray__button, .dice-tray__ad", (event) => {
 				const dataset = event.currentTarget.dataset;
 				const dragData = JSON.parse(JSON.stringify(dataset));
@@ -2016,6 +2243,24 @@ Hooks.on("renderSidebarTab", async (app, html, data) => {
 			}
 		});
 	}
+});
+
+Hooks.on("getSceneControlButtons", (controls) => {
+	const popout = game.settings.get("dice-calculator", "popout");
+	if (popout === "none") return;
+	const autoOpenPopout = game.settings.get("dice-calculator", "autoOpenPopout");
+	const addButton = (control) => {
+		control.tools.push({
+			name: "dice-tray",
+			title: "Dice Tray",
+			icon: "fas fa-dice-d20",
+			onClick: () => togglePopout(),
+			active: CONFIG.DICETRAY.popout?.rendered || (!game.ready && autoOpenPopout),
+			toggle: true,
+		});
+	};
+	if (popout === "tokens") addButton(controls[0]);
+	else controls.forEach((c) => addButton(c));
 });
 
 /**