+++ systems/yzecoriolis/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+### [3.9.1](https://www.github.com/winks-vtt/yze-coriolis/compare/v3.9.0...v3.9.1) (2024-07-21)
+
+
+### Bug Fixes
+
+* can create items in v12 ([300aed9](https://www.github.com/winks-vtt/yze-coriolis/commit/300aed936acde1001a23c97c6635594204812a5b))
+
+## [3.9.0](https://www.github.com/winks-vtt/yze-coriolis/compare/v3.8.0...v3.9.0) (2024-07-20)
+
+
+### Features
+
+* updated support to v12 ([65ecc00](https://www.github.com/winks-vtt/yze-coriolis/commit/65ecc00f2e037fa763aaaa5996f3ae86419d8912))
+
+
+### Miscellaneous
+
+* vsode settings ([44361df](https://www.github.com/winks-vtt/yze-coriolis/commit/44361dfef58d7e3603387a3e72cdc2f9a5a0f3e7))
+
 ## [3.8.0](https://www.github.com/winks-vtt/yze-coriolis/compare/v3.7.0...v3.8.0) (2024-04-22)